Run migrations for each branch.
Rather than running migrations at the beginning (on the first branch)
and at the end (on the final branch), we want to run migrations up and
down on each branch. That allows us to test two branches with migrations
and allows us to properly roll back migrations not in master.